+++ src/pipeline/SecureTcpTransport.cpp
@@ -287,9 +287,10 @@ std::optional<std::string> SecureTcpTransport::readEncryptedLine() {
 
     while (true) {
         if (rawBuffer_.size() >= SECRETBOX_NONCEBYTES + 2) {
-            uint16_t frameLen = (static_cast<uint16_t>(
-                static_cast<unsigned char>(rawBuffer_[SECRETBOX_NONCEBYTES])) << 8)
-                | static_cast<uint16_t>(static_cast<unsigned char>(rawBuffer_[SECRETBOX_NONCEBYTES + 1]));
+            uint16_t frameLen = static_cast<uint16_t>(
+                (static_cast<uint16_t>(
+                    static_cast<unsigned char>(rawBuffer_[SECRETBOX_NONCEBYTES])) << 8)
+                | static_cast<uint16_t>(static_cast<unsigned char>(rawBuffer_[SECRETBOX_NONCEBYTES + 1])));
 
             size_t totalFrame = SECRETBOX_NONCEBYTES + 2 + frameLen;
             if (rawBuffer_.size() >= totalFrame) {
